// SmartChurch.Angular/src/app/modules/people/interfaces/person.interface.ts
export interface IPerson {
  Id: number;
  Name: string;
  Email: string;
  PhoneNumber: number;
  WhatsappNumber: number;
  FacebookLink: string;
  Picture: string;
  Comments: string;
  Birthday: string;
  GradeId: number;
  Job: string;
  MaritalStatusId: number;
  MarriageAnniversary: string;
  Address: string;
  Weekends: number;
  Qualifications: string;
  ReligiousBackgroundId: number;
  CountryOfResidenceId: number;
  HomeCountryId: number;
  IsServant: boolean;
  LastVisitDate: string;
  BaptismCertificationPic: string;
  MarriageCertificatePic: string;
  MarriageTypeId: number;
  LastVisitYearOfMarriageAnniversary: number;
  LastVisitYearOfBirthday: number;
}